msm: camera: isp: Send the current request id for trigger camera
For Trigger camera for rdi only context, state machine needs to start
with epoch, and if in non-early pcr case, need to move the init request
to free list during the stream on time. Added proper trigger logic in the
request manager. send the current request id to request manager

CRs-Fixed: 3118886
Change-Id: Iab5de015d0138ff196c28a879dda591264396844
Signed-off-by: Ravikishore Pampana <quic_rpampana@quicinc.com>