msm: ais: reqmgr: Avoid freeing subdev twice
The 'l_device' pointer in __cam_req_mgr_destroy_subdev is
set to NULL after freeing but this is done on a
local copy of the variable in stack. This results in
double-free when this function is called again. To avoid
this, pass 'l_device' pointer by reference and assign it
to NULL after freeing.

Change-Id: I5a2164e3883d83dcd5ca2a558f1e0f4f20348800
Signed-off-by: E V Ravi <quic_evenka@quicinc.com>