msm: camera: sync: Reset sync object state for fence reuse
UMD can directly use the existing fence after destroy without an actual
ioctl call. The handling is done by CSL, which provides a unique
identifier to the CreateSync's caller function. This uid is stored in
the 16-MSB bits of the sync_obj. We also store this uid as well as
signalled state of the previous fence that used the current sync_obj
so that the callbacks can be handled for the last signalled fence as well.

Change-Id: Ie70730b85199ea00953e4ff6911fa9e099866f39
CRs-Fixed: 3254339
Signed-off-by: Shardul Bankar <quic_sharbank@quicinc.com>