msm: camera: ife: implicit RDI allocation
Each of the IFE multiple RDIs are equivalent but current userspace
interface allows only explicitly requested (by id) RDI to be
allocated. This causes allocation of multiple IFEs, while one can do
the job, and can reduce the functionality of IFE driver in multicamera
usecases. With this change, a check is made if an IFE has another
suitable free interface, if requested is already busy and allocates it
instead. Also communicates with userspace the id of allocated
interface.

Change-Id: I3f38a66641f6fc9278d4da67d8fe5eb58475252f
Signed-off-by: Waseem Akhtar <quic_wakhtar@quicinc.com>