msm: camera: isp: Fix independent crm bayer usecase
Issue:
IFE packet was getting applied again once the packet
is applied, if packet is received from UMD.
This was resulting into bubble if CDM is not able to
apply both packets before epoch.

Fix:
Stop abrupt apply in independent crm case if usecase
is running on pix path.

CRs-Fixed: 3264510
Change-Id: I8ef5a4ea9b97b8c9b041405417b43febfc7515c9
Signed-off-by: Rishabh Jain <quic_risjai@quicinc.com>