msm: ais: fix csid ECC error when do multi-camera start/stop test
If set CSID_CSI2_RX_CFG0[PHY_NUM_SEL] as 0 when stopping camera, then
it effectively gates the stream from PHY/sensor. If you clear this field
before streaming off the sensor on a frame boundary, you potentially cut
off the stream to CSID in the middle: the line(long packet) or a packet
header. Then CSID is unable to complete the CRC/ECC calculations correctly.
There is no need to touch this register when starting and stopping.
So skip reset CSID_CSI2_RX_CFGn registers when stop camera.

Change-Id: I3b6e7205e1cb21edc0557e0f07a08cefad915f3f
Signed-off-by: Liu Jianhua <quic_jianhl@quicinc.com>