msm: camera: sensor: Add reapply setting logic for timeout
Some scenarios due to lack of parallel initilization of sequence
with same CCI master can cause timeout event. In this scenario
initilization setting needs to be reapply for image sensor to be
in expected operational state. This change adds condition to reapply
initilization setting in case of timeout event just one more time.
If initilization settings fails to reapply one more time it is
consider as an error and notify as error case.

Change-Id: I9b7ef3b2f011240c955d19d858b4b9393f51ad44
Signed-off-by: Jigarkumar Zala <quic_jzala@quicinc.com>