msm: camera: isp: Update RUP based on requested ports
If RUP bit is set for a port in first request, and first request
does not have that port enabled, then RUP IRQ will be generated in
second request without applying second request. In this case buf_done
may not be generated in second request.

This can be handles by updating RUP mask to match enbled ports in a
request .

CRs-Fixed: 3386611
Change-Id: I35513be2cdfbc37593ca0ef4d86d1157ba1fe199
Signed-off-by: Trishansh Bhardwaj <quic_tbhardwa@quicinc.com>