msm: camera: isp: Add lazy clear register sanitization
In some hardware, it is possible that global clear command on an
independent controller can clear relevant status bits of a dependent
controller due to the clear register being dirty. This commit
introduces lazy clear register sanitization.
In case of a dirty clear register in a dependent controller, the IRQ
controller waits till the next interrupt and checks if that clear
register is going to get updated. If not, that clear register is
sanitized (set to zero).

CRs-Fixed: 3405951
Change-Id: Ie94252fb378676481410759c8bc87088d27024dd
Signed-off-by: zhuo <quic_zhuo@quicinc.com>