msm: camera: memmgr: Add refcount to track umd in use buffers
Currently krefcount is using by umd and kmd. Due to sometimes
there is issue in umd, such as release twice. That maybe causes
buffer release before kmd access the buffer. This commit add
a new refcount to track umd in use buffers and use current krefcount
to track kmd in use buffers. For the same buffer use in kmd and umd
only when all refcount become zero, the buffer start to release.

CRs-Fixed: 3692103
Change-Id: I513b5de7ef34ca0e0e69f7d039cbe9b46b06a73b
Signed-off-by: zhuo <quic_zhuo@quicinc.com>